Disambiguate mul! for matvec and matmat through indirection (#52837)
This forwards `mul!` for matrix-matrix and matrix-vector multiplications
to the internal `_mul!`, which is then specialized for the various array
types. With this, packages would not encounter any method ambiguity when
they define
```julia
mul!(::AbstractMatrix, ::MyMatrix, ::AbstractMatrix, alpha::Number, beta::Number)
```
This should reduce the number of methods that packages need to define to
work around ambiguities with `LinearAlgebra`.

There was already an existing internal function named `_mul!`, but the
new methods don't clash with the existing ones, and since both sets of
methods usually forward structured multiplications to specialized
functions, it's fitting for them to have the same name.