Add upper/lowertriangular functions and use in applytri (#53573)
We may use the fact that a `Diagonal` is already triangular to avoid
adding a wrapper.

Fixes the specific example in
https://github.com/JuliaLang/julia/issues/53564, although not the
broader issue. This is because it changes the operation from a
`UpperTriangular + UpperTriangular` to a `UpperTriangular + Diagonal`,
which uses broadcasting. The latter operation may also allow one to
define more efficient methods.