Targeted `stable_muladdmul` branches in small matmul (#54470)
This PR changes the signature of `matmul2x2!` and `matmul3x3!` so that
they now accept the alpha and beta parameters instead of the
`MulAddMul`. Consequently, the `@stable_muladdmul` macro may be applied
to smaller blocks of code within the functions instead of applying it to
the entire call. This reduces the matrix multiplication latency by a
bit.

```julia
julia> using LinearAlgebra

julia> A = rand(2,2); B = rand(size(A)...); C = zeros(size(A));

julia> @time mul!(C, A, B);
  0.903632 seconds (2.54 M allocations: 130.322 MiB, 5.52% gc time, 100.00% compilation time) # master
  0.813335 seconds (2.28 M allocations: 117.025 MiB, 2.07% gc time, 100.00% compilation time) # This PR
```