LAPACK: Avoid repr call in `chkvalidparam` (#54952)
We were calling `repr` here to interpolate the character with the quotes
into the error message. However, this is overkill for this application,
and `repr` introduces dynamic dispatch into the call. This PR hard-codes
the quotes into the string, which matches the pattern followed in the
other error messages following `chkvalidparam`.