Update help info for qr.jl (#55320)
'Rather than' seems to be the correct intent of the help info. This
would mean that Julia stores the result of QR in compact form instead of
(=rather than) storing in dense form.

'Rather as' would mean that Julia stores Q and R in two separate dense
matrices and not in compact form.