faster implementation of rank(::QRPivoted) fixes #1128 (#1129)
(fixes #1128)
The current implementation of `rank(::QRPivoted)` is slow and inefficient because of repeated calls to `A.R`, each of which triggers memory allocation. This PR fixes that by using `A.factors` instead, which does not cause memory allocations. The proposed implementation is orders of magnitude faster and avoids unnecessary memory allocations altogether.