Fallback `newindex` method with a `BandIndex` (#1143)
Currently, `Broadcast.newindex` with a `BandIndex` is specialized for
structured broadcasting, but this incorrectly assumed that all the
matrices involved in a structured broadcast operation will be structured
matrices. This PR adds a fallback that simply converts the `BandIndex`
to a `CartesianIndex` before computing the `newindex`.

This PR gets the tests for `FillArrays` working on julia nightly.